+++ src/Modules/Payments/Explorer.Payments.Core/Domain/Wallet.cs
@@ -12,8 +12,8 @@ namespace Explorer.Payments.Core.Domain
         public long UserId { get; init; }
 
         public string Username { get; init; }
-        public int AC { get; set; }
-        public Wallet(long userId, string username, int aC)
+        public double AC { get; set; }
+        public Wallet(long userId, string username, double aC)
         {
             UserId = userId;
             Username = username;
